requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
tabulate>=0.8.9
configparser>=5.0.0